64bit cleanup, work around for Debian/Ubuntu amd64, do not use deprecated pattern %_
git-svn-id: file:///tmp/crunc/trunk/perl/ext/Language-Prolog-Yaswi@1574 5310da26-23bb-0310-a32f-9ca5ae3db751

+++ Low/hook.c
@@ -19,7 +19,7 @@ int hook_set=0;
 static int my_agc_hook(atom_t a) {
     if(!strcmp(OPAQUE_FUNCTOR, PL_atom_chars(a))) {
         const char *oname;
-        int olen;
+        size_t olen;
 	dTHX;
 	dSP;
 	ENTER;
